scanner: rewrite directory traversal to work with two passes
the first pass is now exclusively about i/o: scanning the directory to generate
a tree of translation units in directories. the second pass takes the tree and
performs a scan to create targets out of it.

this simplifies things a little and makes the responsibility seperation clear.

at the same time, we also now have target creation working to spec.